+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### HEAD
 
+* Remove grunt-contrib-clean
 * Add node-matchdep to automatically fetch dependencies
 * Add clean Grunt config at beginning of Gruntfile (#46)
 * Use only one Apple Touch Icon